Fixing reversibility issues

+++ experiments/9_training_stack_optimisation_and_cost_governor/training/deepspeed_template/src/models/model_3b.py
@@ -1031,7 +1031,7 @@ class MoEFFN(nn.Module):
                 sorted_out[start:end] = h @ self.W_down[e]
             start = end
 
-        weighted_out = sorted_out * sorted_weights.unsqueeze(-1)
+        weighted_out = sorted_out * sorted_weights.unsqueeze(-1).to(dtype)
         routed_out = torch.zeros(N, D, device=device, dtype=dtype)
         routed_out.scatter_add_(0, sorted_token_indices.unsqueeze(-1).expand(-1, D), weighted_out)
